feat(client): add Agent and Skill pages
