+++ packages/client/src/components/shared/app-sidebar.tsx
@@ -3,10 +3,12 @@ import {
 	Bolt,
 	Bot,
 	CircleUserRound,
+	Computer,
 	Database,
 	Home,
 	LayoutGrid,
 	PanelLeftOpen,
+	Puzzle,
 	Settings,
 	ShieldCheck,
 	Sigma,
@@ -40,6 +42,16 @@ const CATALOG_ROUTES = [
 		text: "Apps",
 		icon: <LayoutGrid className="size-4" />,
 		route: "/app",
+	},
+	{
+		text: "Agents",
+		icon: <Computer className="size-4" />,
+		route: "/agent",
+	},
+	{
+		text: "Skills",
+		icon: <Puzzle className="size-4" />,
+		route: "/skill",
 	},
 	{
 		text: "Model",
