+++ packages/client/src/pages/agent/agent-page.tsx
@@ -1,0 +1,3 @@
+import { CatalogPage } from "../app/catalog-page";
+
+export const AgentPage = () => <CatalogPage type="agent" />;
