+++ packages/client/src/pages/skill/skill-page.tsx
@@ -1,0 +1,3 @@
+import { CatalogPage } from "../app/catalog-page";
+
+export const SkillPage = () => <CatalogPage type="skill" />;
